Adds links, images shortcuts, void support
Adds the links and images methods to ElementsJS, which let you create collections of links and images for easier reference in code.

Also adds the void property, which is checked when shortcuts are initialized. When void is true, the shortcut returned is the void element, rather than a function that accepts a template string.