Fixes shortcuts, adds parentSelector
This fixes a bug where the void elements were added to new Shortcuts as elements, rather than a function that would return an element. Now, the void property should work as expected for shortcuts.

Additionally, this commit adds the parentSelector property to elements. This functions similarly to parentNode (the properties are mutually exclusive), but it accepts a selector rather than a node.