Update README.md
Incremental update of documentation.

Negative behavior identified, too: Right now, because of the way I'm applying text when the _e template string is used, the content of the element isn't added to the template. A couple of ways this could be handled: the text could be added to the template before the element is created, or we could use updateJSONtemplate to add the text after the node exists. Right now, handling everything via the template is the preferred solution to implement.